Moved config of firebase to config folder
